Updated about-us page into components
